requirements.txt
# Core
anthropic>=0.25.0
python-dotenv>=0.19.1

# Web server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Twitter API (posting via official API)
tweepy>=4.15.0

# Browser automation — Playwright ships its own Chromium, no system browser needed
playwright>=1.40.0

# Data
pandas>=2.0.0
fake-headers>=1.0.2

# UI
rich>=13.9.4

# WebSocket
websockets>=12.0
aiofiles

# OpenAI (fine-tuned brain)
openai>=1.0.0

httpx>=0.27.0
requests-oauthlib>=1.3.1
itsdangerous>=2.1.2

# Cortical Labs CL SDK — neural substrate integration
cl-sdk>=0.1.0
h5py>=3.9.0
tables>=3.9.0
scipy>=1.11.0
msgpack>=1.0.5
msgpack-numpy>=0.4.8
solders>=0.21.0
